Move two misplaced src/ subjects to fitting homes (#61944)
