+++ src/workflows/sync-sdk-docs/convert-mermaid.ts
