+++ src/workflows/sync-sdk-docs/normalize-sdk-docs.ts
